refactor(database): move database decompression headers under decompression dir
