+++ includes/acl/decompression/decompress.h
@@ -34,8 +34,8 @@
 #include "acl/core/track_traits.h"
 #include "acl/core/track_types.h"
 #include "acl/core/impl/compiler_utils.h"
-#include "acl/database/database.h"
 #include "acl/decompression/decompression_settings.h"
+#include "acl/decompression/database/database.h"
 #include "acl/decompression/impl/decompression_context_selector.h"
 #include "acl/decompression/impl/decompression_version_selector.h"
 #include "acl/decompression/impl/scalar_track_decompression.h"
