+++ includes/acl/decompression/impl/decompression_version_selector.h
@@ -29,7 +29,7 @@
 #include "acl/core/error.h"
 #include "acl/core/interpolation_utils.h"
 #include "acl/core/impl/compiler_utils.h"
-#include "acl/database/database.h"
+#include "acl/decompression/database/database.h"
 #include "acl/decompression/impl/scalar_track_decompression.h"
 #include "acl/decompression/impl/transform_track_decompression.h"
 #include "acl/decompression/impl/universal_track_decompression.h"
