+++ includes/acl/decompression/impl/scalar_track_decompression.h
@@ -30,7 +30,7 @@
 #include "acl/core/track_writer.h"
 #include "acl/core/variable_bit_rates.h"
 #include "acl/core/impl/compiler_utils.h"
-#include "acl/database/database.h"
+#include "acl/decompression/database/database.h"
 #include "acl/math/scalar_packing.h"
 #include "acl/math/vector4_packing.h"
