+++ tools/acl_compressor/sources/acl_compressor.cpp
@@ -53,10 +53,10 @@
 #include "acl/compression/transform_pose_utils.h"	// Just to test compilation
 #include "acl/compression/impl/write_decompression_stats.h"
 #include "acl/compression/track_error.h"
-#include "acl/database/database.h"
-#include "acl/database/null_database_streamer.h"	// Just to test compilation
-#include "acl/database/impl/debug_database_streamer.h"
 #include "acl/decompression/decompress.h"
+#include "acl/decompression/database/database.h"
+#include "acl/decompression/database/null_database_streamer.h"	// Just to test compilation
+#include "acl/decompression/database/impl/debug_database_streamer.h"
 #include "acl/io/clip_reader.h"
 
 #include <cstring>
